requirements.txt
coremltools==5.2.0
Flask==2.2.2
keras==2.9.0
matplotlib==3.5.2
numpy==1.22.4
onnx==1.12.0
opencv_python==4.6.0.66
pafy==0.5.5
pandas==1.4.3
Pillow==9.2.0
requests==2.28.1
scipy==1.9.0
seaborn==0.11.2
setuptools==62.6.0
thop==0.1.1.post2207130030
torch==1.12.0+cu116
torchvision==0.13.0+cu116
tqdm==4.64.0
wandb==0.13.2
